Add post update to dom
