+++ src/components/App/App.js
@@ -1,12 +1,13 @@
 import React, { useState, useEffect } from 'react';
 import './App.css';
-import { getUrls } from '../../apiCalls';
+import { getUrls, postUrl } from '../../apiCalls';
 import UrlContainer from '../UrlContainer/UrlContainer';
 import UrlForm from '../UrlForm/UrlForm';
 
 const App = () => {
   const [urls, setUrls] = useState([])
   const [appError, setAppError] = useState()
+  const [message, setMessage] = useState()
 
   const fetchUrls = async () => {
     setAppError('')
@@ -14,6 +15,24 @@ const App = () => {
     try {
       const data = await getUrls()
       setUrls(data.urls)
+    } catch (error) {
+      setAppError(error.message)
+    }
+  }
+
+  const sendInfo = async (title, url) => {
+    setAppError('')
+    setMessage('')
+
+    const body = {
+      "long_url": url,
+      "title": title
+    }
+
+    try {
+      const post = await postUrl(body)
+      setUrls([...urls, post])
+      setMessage(`Url successfully shortended to ${post.short_url}!`)
     } catch (error) {
       setAppError(error.message)
     }
@@ -27,7 +46,7 @@ const App = () => {
       <main className="App">
         <header>
           <h1>URL Shortener</h1>
-          <UrlForm />
+          <UrlForm message={message} appError={appError} sendInfo={sendInfo}/>
         </header>
         {appError && <p>{appError}</p>}
         <UrlContainer urls={urls}/>
